Создана директория промежуточных слоев. Перенесен промежуточный слой авторизации в новую директорию.
